Create List creation and inverse sorting
this is for homework that I tried to make more complicated to better reflect the things I can do. it makes a either random or self input list that. It then inversely sorts (big numbers first before small numbers). Last it prints off each element of the list from largest to smallest.